Добавил Поле http://localhost:8080

+++ OperationsService/src/main/java/com/murad/operationsservice/service/BuyService.java
@@ -91,15 +91,16 @@ public class BuyService {
             buyRepository.save(buyEntity);
         }
         ///////////////
-        if (buyEntities != null && buyEntities.size() > 0) {
-            long userId = buyEntities.get(0).getUserId();
-            List<Long> productIds = buyEntities.stream().map(buyEntity -> buyEntity.getProductId()).toList();
-            FromCartDeleteDto fromCartDeleteDto = FromCartDeleteDto.builder()
-                    .userId(userId)
-                    .productIds(productIds).build();
-            cartServiceClient.deleteAllFromCart(fromCartDeleteDto);
+        if (!dto.isError()) {
+            if (buyEntities != null && buyEntities.size() > 0) {
+                long userId = buyEntities.get(0).getUserId();
+                List<Long> productIds = buyEntities.stream().map(buyEntity -> buyEntity.getProductId()).toList();
+                FromCartDeleteDto fromCartDeleteDto = FromCartDeleteDto.builder()
+                        .userId(userId)
+                        .productIds(productIds).build();
+                cartServiceClient.deleteAllFromCart(fromCartDeleteDto);
+            }
         }
-
     }
 
     public PaymentRequest buyProduct(BuyProductDto buyProductDto) {
